requirements.txt
# Core
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
psycopg2-binary==2.9.10

# Redis
redis[hiredis]==5.2.1

# AI / LLM
groq==0.12.0
langchain==0.3.10
langchain-groq==0.2.3
langgraph==0.2.59
langchain-community==0.3.10

# RAG / Embeddings
chromadb==0.5.23
sentence-transformers==3.3.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20

# Utilities
httpx==0.28.1
aiofiles==24.1.0
orjson==3.10.12
structlog==24.4.0
tenacity==9.0.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
